@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: '>=2.19.2 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,6 +33,7 @@ dependencies:
   bloc: ^8.1.1
   flutter_bloc: ^8.1.2
   dio: ^5.1.1
+  match: ^0.4.1
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -42,6 +43,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  match_generator: ^0.4.1
+  build_runner: ^2.3.3
 
 
   # The "flutter_lints" package below contains a set of recommended lints to
