@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:register_of_receivables_front/pages/people_list/people_list_page.dart';
 
 import 'core/core.dart';
 import 'pages/pages.dart';
@@ -9,7 +8,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -20,7 +19,7 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(primarySwatch: Colors.blue),
         initialRoute: '/people_list',
         routes: {
-          '/people_list': (context) => PeopleListPage(),
+          '/people_list': (context) => PeopleListRouter.page,
           '/people_form': (context) => PeopleFormPage(),
         },
       ),
